// frontend/src/api-client/model/questionRequestDTO.ts
/**
 * Workshop 2024 MKL
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface QuestionRequestDTO { 
    title: string;
    a1: string;
    a2: string;
    a3: string;
    a4: string;
    correct: number;
    sortIdx: number;
}