// frontend/src/api-client/model/quizRequestDTO.ts
/**
 * Workshop 2024 MKL
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QuestionRequestDTO } from './questionRequestDTO';


export interface QuizRequestDTO { 
    quizName: string;
    questions: Array<QuestionRequestDTO>;
}