// frontend/src/api-client/model/quizResponseDTO.ts
/**
 * Workshop 2024 MKL
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QuestionResponseDTO } from './questionResponseDTO';


export interface QuizResponseDTO { 
    id: string;
    createdAt: string;
    quizName: string;
    questions: Array<QuestionResponseDTO>;
}