// frontend/src/api-client/model/quizUpdateRequestDTO.ts
/**
 * Workshop 2024 MKL
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QuestionUpdateRequestDTO } from './questionUpdateRequestDTO';


export interface QuizUpdateRequestDTO { 
    id: string;
    quizName: string;
    questions: Array<QuestionUpdateRequestDTO>;
}